requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
dj-email-url==1.0.6
Django==4.2.2
django-cache-url==3.4.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.19.0
packaging==23.1
phonenumbers==8.13.15
psycopg2-binary==2.9.6
pycountry==22.3.5
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
whitenoise==6.5.0
